+++ cayleypy/hasher.py
@@ -55,22 +55,26 @@ class StateHasher:
 
     def _make_hashes_cpu_and_modern_gpu(self, states: torch.Tensor) -> torch.Tensor:
         if states.shape[0] <= self.chunk_size:
-            return (states @ self.vec_hasher).reshape(-1)
+            return (states.to(torch.int64) @ self.vec_hasher).reshape(-1)
         else:
             parts = int(math.ceil(states.shape[0] / self.chunk_size))
-            return torch.vstack([z @ self.vec_hasher for z in torch.tensor_split(states, parts)]).reshape(-1)
+            return torch.vstack(
+                [z.to(torch.int64) @ self.vec_hasher for z in torch.tensor_split(states, parts)]
+            ).reshape(-1)
 
     def _make_hashes_older_gpu(self, states: torch.Tensor) -> torch.Tensor:
         if states.shape[0] <= self.chunk_size:
-            return torch.sum(states * self.vec_hasher, dim=1)
+            return torch.sum(states.to(torch.int64) * self.vec_hasher, dim=1)
         else:
             parts = int(math.ceil(states.shape[0] / self.chunk_size))
-            return torch.hstack([torch.sum(z * self.vec_hasher, dim=1) for z in torch.tensor_split(states, parts)])
+            return torch.hstack(
+                [torch.sum(z.to(torch.int64) * self.vec_hasher, dim=1) for z in torch.tensor_split(states, parts)]
+            )
 
     def _hash_splitmix64(self, x: torch.Tensor) -> torch.Tensor:
         n, m = x.shape
         h = torch.full((n,), self.seed, dtype=torch.int64, device=x.device)
         for i in range(m):
-            h ^= _splitmix64(x[:, i])
+            h ^= _splitmix64(x[:, i].to(torch.int64))
             h = h * 0x85EBCA6B
         return h
